requirements.txt
pyTelegramBotAPI
mega.py